Update SlideNavigationController with helper function - removeAllMenuFromSuperview
An issue a client of mine faced was with loading interstitial ads on top of the root controller.
The current library doesn't present a way to clean up the menu from the superview.
This function is only invoked when we switch from the left -> right menu, or vice versa.
Provided a public method that can be invoked in such circumstances that will clean up either the left or right menu (whichever is active) as well as reset the lastRevealedMenu